requirements.txt
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
Authlib==1.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==24.2.0
bleach==6.1.0
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
cryptography==42.0.4
cssselect==1.2.0
cssutils==2.9.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
docopt==0.6.2
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
firebase-admin==6.4.0
flake8==7.0.0
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
functions-framework==3.5.0
fuzzywuzzy==0.18.0
google==3.0.0
google-api-core==2.17.0
google-api-python-client==2.121.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-tasks==2.16.2
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gotrue==2.1.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.25.2
idna==3.6
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
msgpack==1.0.7
mypy==1.8.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
oauthlib==3.2.2
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
pluggy==1.4.0
postgrest==0.15.0
premailer==3.10.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pyzmq==25.1.2
realtime==1.0.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.5
supafunc==0.3.3
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
wheel==0.41.2
yagmail==0.15.293
yarg==0.1.9